fix: realign autopkgtest libatomic check with the verifier
The autopkgtest to check Rumur could generate a compilable model contains logic
to decide whether linking against libatomic is needed, but this logic did not
correspond to the actual verifier code closely enough. The __sync built-in it
tests can be lowered to a single instruction on ≥ armv8.1-a, but the __atomic
built-ins used in the actual verifier cannot.

Changes to the verifier itself should now allow it to be lock-free on ≥
armv8.1-a. This change updates the logic used to check libatomic requirements to
something that approximates the new verifier.

Debian: fixes https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=1018205
Reported-by: Tobias Frost <tobi@debian.org>
Thanks: Caroline Xu