use close-on-exec to more simply clean up pipe file descriptors
A recent bug fix in the Zig compiler¹ made me realise this code was
unnecessarily carrying open file descriptors (the source ones to the `dup2`
operations) into the child process. We can address this and simplify the whole
situation by taking advantage of the `dup2` happening _before_ exec, and thus
close-on-exec does what we need.

¹ https://github.com/ziglang/zig/commit/389181f6be8810b5cd432e236a962229257a5b59