fix: allow writing 'VALUE_MIN' naturally
When writing the natural form of the minimum of a signed `value_t`, Rumur would
abort at runtime claiming this was an out-of-range value. The problem is that
such a thing was lowered into negation of an out-of-range value. This change
avoids this with the same trick libcs use to define `INT_MIN`.