support beginning parsing with a fragment, rather than a full model
For the companion tools to Rumur like `murphi2uclid`, it is becoming desirable
to translate fragments of Murphi syntax without requiring a fully formed model.
This is a step towards allowing this. None of this functionality is actually
exposed yet (`parse` hard codes `START_MODEL`).

The supported fragments are a declaration, expression, property, rule, or
statement. Restricting to these rather than every single parsing rule maps
roughly to what the user might think of as comprehensible units and avoids
ambiguity. That is, a future planned API that performs automatic discrimination
of what type of entity is being parsed has some chance of giving the answer the
user expects. Without out this, e.g. it is ambiguous whether `42` is an
expression or a number.

The implementation of this looks like a hack, but is apparently the
Bison-recommended way of achieving this.¹ Adapting the recipe for this to work
when generating a C++ parser was a bit of brain teaser, but this seems to be the
solution.

¹ https://www.gnu.org/software/bison/manual/html_node/Multiple-start_002dsymbols.html