fix: reject non-boolean expressions in properties
It was possible to write a property with an expression that was not boolean and
have this pass through Rumur successfully. The generated code would not type
check and would fail compilation.

Surprisingly this bug has been resident for years and only discovered now by
Halfempty,¹ a test case reducer, while debugging an orthogonal issue.

¹ https://github.com/googleprojectzero/halfempty

Github: #277 “'-Wmaybe-uninitialized' warnings”
Reported-by: 10f7a122b938d0d41f91bcdbe412c828c00519aa