fix: avoid -Wmaybe-uninitialized false positives
When returning an aggregate type from a function, Rumur would generate a byte
buffer as a local that was not initialised. This was functionally correct, as
data would be written into the buffer before it was returned. However these
writes occur through functions like `handle_copy` that have sufficient
dynamism/generality that the compiler cannot tell they are not reading
uninitialised parts of the buffer. As a result, the compiler would emit
`-Wmaybe-uninitialized` warnings when compiling the generated code.

We could try to work around this in various ways, but the simplest solution
seems to be to just initialise the buffer, making it clear to humans and the
compiler that no reads of uninitialised data are happening. This is consistent
with how aggregate local variables and parameters are handled.

Github: closes #277 “'-Wmaybe-uninitialized' warnings”
Reported-by: 10f7a122b938d0d41f91bcdbe412c828c00519aa