CI: move macOS jobs to faster image
Commit 916f8a877748b517c4f5537ad5900167861a1b11 of the Cirrus CI docs¹ removed
mention of the images we were using. The commit itself is fairly cryptic and its
originating pull request does not explain further either. But following through
one of the URLs removed in that commit and clicking a promising sounding link
leads to a blog post² that explains these new images save ~10s of boot up time
due to keeping hot VMs on standby instead of cold booting each time.

¹ https://github.com/cirruslabs/cirrus-ci-docs/commit/916f8a877748b517c4f5537ad5900167861a1b11
² https://cirrus-runners.app/blog/2024/04/11/optimizing-startup-time-of-cirrus-runners