port test suite to Pytest
Traditionally, Rumur has limited itself to only modules present in the Python
standard library, so as not to force users/developers of something that is
primarily a C/C++ program to deal with Python package installation. This has
necessitated hand rolling test suite mechanisms (e.g. parametrisation) and doing
without some bells and whistles (e.g. xfail). The costs of this bespoke setup
have begun to outweigh their utility. Hopefully, in 2025, taking a dependency on
Pytest is acceptable.

This change aims to preserve compatibility with Python 3.4.