port make-version.py to 'pathlib.Path'
A surprising subtlety in this change is that the absolute-ising of the paths
needs to be retained. On some platforms (e.g. GCC 8.5 image with Python 3.7)
this script is run from within librumur/ (maybe a CMake 3.13.4 behaviour?). This
means it is not possible to `.parents[2]` because the relative path does not
have that many levels.