update librumur-api Autopkgtest
603a2e4bd4a8ef3a99907f408a04e361dbe9f3bb updated the API in a breaking way that
requires changing this code. Although this change happened a couple of releases
back, there has been no Debian packaging in the meantime so this was not
realised when scanning for changes since the last release. The fact that this
was not caught pre-RFS indicates a bit of a hole in our process (Autopkgtests
are apparently not run) which should be addressed in future.

Debian: https://lists.debian.org/debian-mentors/2025/01/msg00047.html
Reported-by: Phil Wyett <philip.wyett@kathenas.org>