fix: locate Pytest cache within the build directory
When the test suite was run, this would produce a .pytest_cache directory in the
source directory. This erroneously causes test runs from different builds to
share a test cache.

This was detected during Debian packaging and an attempt at suppression was made
in 729a325da1e78e5d306e4a124a38783089c6fe55. But this was insufficient for
validation during the Debian RFS process. What went unrecognised when
constructing the suppression was that this is not merely an annoyance but
actually functionally incorrect in the presence of multiple builds using the
same source directory.

Debian: https://lists.debian.org/debian-mentors/2025/01/msg00047.html
Reported-by: Phil Wyett <philip.wyett@kathenas.org>