invert mode of interaction with the standard library
The standard library list template was written in the style of an X macro.¹ This
was purely because I was not creative enough to realise we could define an M4
macro that would expand to the required Murphi. Making this a macro definition
makes interaction with it much more natural, as well as not requiring the list
file to be included multiple times.

¹ https://en.wikipedia.org/wiki/X_macro