fix malformed generated code when complex-returning function used within guard
When the return value of a complex-returning function was used within a guard,
it would be emitted as code referring to preceding declared backing storage.
This was and is correct. However the backing storage it refers to was never
itself being defined.

While this was a simple omission, the cause of this seems to have been an
oversight in my own thinking. I was not creative enough to realise that it was
possible to use a complex-returning function in this context.

Github: fixes #290 “error when chaining a function returning a record type
  within a rule guard”
Reported-by: Julian Pritzi