use atomics when accessing queue elements
Traditionally non-atomic accesses here have been fine (I believe) because the
`__ATOMIC_SEQ_CST` reads and writes of the `queue_node`s effectively creates a
total order of the accesses to the contained queue elements. However upcoming
changes will relax the `queue_node` accesses in a way that I believe then
exposes the queue element accesses to races. This change should be enough to
avoid this.